+++ psutil/tests/test_contracts.py
@@ -181,9 +181,17 @@ class TestDeprecations(unittest.TestCase):
         with warnings.catch_warnings(record=True) as ws:
             psutil.Process().memory_info_ex()
         w = ws[0]
-        self.assertIsInstance(w.category(), FutureWarning)
+        self.assertIsInstance(w.category(), DeprecationWarning)
         self.assertIn("memory_info_ex() is deprecated", str(w.message))
         self.assertIn("use memory_info() instead", str(w.message))
+
+    @unittest.skipIf(not MACOS, "deprecated on macOS")
+    def test_memory_maps_osx(self):
+        with warnings.catch_warnings(record=True) as ws:
+            with self.assertRaises(psutil.AccessDenied):
+                psutil.Process().memory_maps()
+        w = ws[0]
+        self.assertIsInstance(w.category(), DeprecationWarning)
 
 
 # ===================================================================